♻️ refactor(chat): organize agentRun module into layered subfolders
Phase B of the agent-run module layout refactor (behavior-preserving moves;
no logic changes).

Reorganize the flat `agentRun/actions/` (14 files) into the layers it had
already grown:

- dispatch/        selectRuntimeType + sub-agent dispatch (routing, no effects)
- lifecycle/       run-lifecycle control plane (buildRunLifecycle, types) +
                   the client signal bridge
- entries/         sendMessage/compression (conversationLifecycle), intervention
                   (conversationControl), commandBus
- transports/      one folder per runtime: client / gateway / hetero
- state/           streamingStates, memory

Imports updated repo-wide; depth-sensitive relative imports in the moved
files normalized to `@/` absolute paths.
